Convert Integer to Long values

+++ src/main/java/com/lickhunter/spotbot/controllers/ApplicationController.java
@@ -47,8 +47,8 @@ public class ApplicationController {
                     .forEach(datum -> symbolRepository.save(symbolConverter
                                 .convertFromDto(new SymbolDto()
                                         .withSymbol(datum.getS().concat("USDT"))
-                                        .withGalaxyScore(datum.getGs())
-                                        .withAltRank(datum.getAcr())
+                                        .withGalaxyScore(datum.getGs().intValue())
+                                        .withAltRank(datum.getAcr().intValue())
                                 )));
         }
         return ResponseEntity.ok(null);
